recovery_tools: support Security Version Number (SVN)
Use the version attribute of the XML to emulate the SVN only for ROT type.
Please note, both bmc and pch svn, major and minor version are set in pfm.
The version attribute as below.

Version is 00.01.06 (SVN:00, Major: 01, Minor:06)
```
<RecoveryImage version="00.01.06" platform="rot" type="2">
```

By default, the version string is in ASCII code data, changes the data format to hex data. SVN should be between 00-64 and the version length is 6.
rot version structure:
```
struct rot_version {
   uint8_t svn;
   uint8_t reserved1;
   uint8_t major;
   uint8_t reserved2;
   uint8_t minor;
   uint8_t reserved3;
};
```

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I009201dcdf1310b1e9c9b4328acd1c7cc373bc16